+++ link-move/src/main/java/com/nhl/link/move/runtime/connect/IConnectorFactory.java
@@ -3,22 +3,28 @@ package com.nhl.link.move.runtime.connect;
 import com.nhl.link.move.LmRuntimeException;
 import com.nhl.link.move.connect.Connector;
 
+import java.util.Optional;
+
 /**
  * A factory to create connectors of a related type.
- * 
- * @since 1.1
  *
- * @param <C>
- *            a broad type of connector handled by this factory. It shouldn't
+ * @param <C> a broad type of connector handled by this factory. It shouldn't
  *            have to be the exact class of connector, but perhaps a
  *            subinterface like JdbcConnector, LdapConnector, etc.
+ * @since 1.1
  */
 public interface IConnectorFactory<C extends Connector> {
 
-	/**
-	 * Creates a connector for a given symbolic ID. If a connector can not be
-	 * created, or an ID is not recognized, {@link LmRuntimeException} is
-	 * thrown.
-	 */
-	C createConnector(String id);
+    /**
+     * @since 3.0
+     */
+    Class<C> getConnectorType();
+
+    /**
+     * Creates a connector for a given symbolic ID. If an ID is not recognized, an empty Optional is returned. If a
+     * connector can not be created, {@link LmRuntimeException} is thrown.
+     *
+     * @since 3.0
+     */
+    Optional<? extends C> createConnector(String id) throws LmRuntimeException;
 }