+++ link-move/src/main/java/com/nhl/link/move/runtime/connect/URIConnectorFactory.java
@@ -7,35 +7,47 @@ import com.nhl.link.move.connect.URIConnector;
 import java.net.URI;
 import java.net.URISyntaxException;
 import java.net.URL;
+import java.util.Optional;
 
 /**
  * @since 1.4
  */
+// TODO: deprecate this factory. We should not treat connector IDs as meaningful URLs
 public class URIConnectorFactory implements IConnectorFactory<StreamConnector> {
 
-	@Override
-	public StreamConnector createConnector(String id) {
+    @Override
+    public Class<StreamConnector> getConnectorType() {
+        return StreamConnector.class;
+    }
 
-		URI uri;
-		try {
-			// first try to get resource from classpath
-			URL url = URIConnectorFactory.class.getResource(id);
-			if (url != null) {
-				uri = url.toURI();
-			} else {
-				// probably we have absolute path to file here
-				uri = new URI(id);
-			}
-			if (!uri.isAbsolute()) {
-				throw new LmRuntimeException(
-						"Provided URI is not absolute (should be a classpath resource or absolute path to file)"
-				);
-			}
-		} catch (URISyntaxException e) {
-			throw new LmRuntimeException("Failed to create connector for URI, because it is malformed: " + id, e);
-		}
+    @Override
+    public Optional<URIConnector> createConnector(String id) {
+        URI uri = toUri(id);
+        return Optional.ofNullable(uri)
+                .map(this::checkAbsolute)
+                .map(URIConnector::new);
+    }
 
-		return new URIConnector(uri);
-	}
+    private URI toUri(String id) {
+        try {
+            // first try to get resource from classpath
+            URL url = URIConnectorFactory.class.getResource(id);
+            if (url != null) {
+                return url.toURI();
+            } else {
+                // probably we have absolute path to file here
+                return new URI(id);
+            }
+        } catch (URISyntaxException e) {
+            return null;
+        }
+    }
 
+    private URI checkAbsolute(URI uri) {
+        if (!uri.isAbsolute()) {
+            throw new LmRuntimeException("Provided URI is not absolute (should be a classpath resource or absolute path to file)");
+        }
+
+        return uri;
+    }
 }