+++ link-move/src/main/java/com/nhl/link/move/runtime/jdbc/DataSourceConnectorFactory.java
@@ -1,39 +1,37 @@
 package com.nhl.link.move.runtime.jdbc;
 
-import com.nhl.link.move.LmRuntimeException;
 import com.nhl.link.move.runtime.connect.IConnectorFactory;
 
 import javax.sql.DataSource;
 import java.util.Map;
+import java.util.Optional;
 
 /**
  * A factory for {@link JdbcConnector} variety of Connectors based on a map of
  * DataSources. Keys in the map correspond to connector IDs.
- * 
+ *
  * @since 1.1
  */
 public class DataSourceConnectorFactory implements IConnectorFactory<JdbcConnector> {
 
-	private final Map<String, DataSource> dataSources;
+    private final Map<String, DataSource> dataSources;
 
-	public DataSourceConnectorFactory(Map<String, DataSource> dataSources) {
-		this.dataSources = dataSources;
-	}
+    public DataSourceConnectorFactory(Map<String, DataSource> dataSources) {
+        this.dataSources = dataSources;
+    }
 
-	@Override
-	public JdbcConnector createConnector(String id) {
-		return new DataSourceConnector(id, connectorDataSource(id));
-	}
+    @Override
+    public Class<JdbcConnector> getConnectorType() {
+        return JdbcConnector.class;
+    }
 
-	DataSource connectorDataSource(String id) {
+    @Override
+    public Optional<JdbcConnector> createConnector(String id) {
+        return connectorDataSource(id).map(ds -> new DataSourceConnector(id, ds));
+    }
 
-		DataSource ds = dataSources.get(id);
-
-		if (ds == null) {
-			throw new LmRuntimeException("Unknown connector ID: " + id + "; available IDs: " + dataSources.keySet());
-		}
-
-		return ds;
-	}
-
+    Optional<DataSource> connectorDataSource(String id) {
+        DataSource ds = dataSources.get(id);
+        return Optional.ofNullable(ds);
+    }
 }