+++ link-move/src/test/java/com/nhl/link/move/unit/LmIntegrationTest.java
@@ -1,12 +1,11 @@
 package com.nhl.link.move.unit;
 
 import com.nhl.link.move.Execution;
-import com.nhl.link.move.connect.Connector;
-import com.nhl.link.move.connect.StreamConnector;
 import com.nhl.link.move.runtime.LmRuntime;
 import com.nhl.link.move.runtime.LmRuntimeBuilder;
 import com.nhl.link.move.runtime.connect.URIConnectorFactory;
 import com.nhl.link.move.runtime.jdbc.DataSourceConnector;
+import com.nhl.link.move.runtime.jdbc.JdbcConnector;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 
@@ -27,11 +26,11 @@ public abstract class LmIntegrationTest extends DerbySrcTargetTest {
     }
 
     protected LmRuntimeBuilder testRuntimeBuilder() {
-        Connector c = new DataSourceConnector("derbysrc", srcDb.getDataSource());
+        JdbcConnector c = new DataSourceConnector("derbysrc", srcDb.getDataSource());
         return new LmRuntimeBuilder()
                 .withTargetRuntime(targetCayenne.getRuntime())
-                .withConnector("derbysrc", c)
-                .withConnectorFactory(StreamConnector.class, new URIConnectorFactory());
+                .connector(JdbcConnector.class, "derbysrc", c)
+                .connectorFactory(new URIConnectorFactory());
     }
 
     protected void assertExec(int extracted, int created, int updated, int deleted, Execution exec) {